move compiler to a seperate module
