+++ qutip/qip/compiler/gatecompiler.py
@@ -1,0 +1,117 @@
+# This file is part of QuTiP: Quantum Toolbox in Python.
+#
+#    Copyright (c) 2011 and later, Paul D. Nation and Robert J. Johansson.
+#    All rights reserved.
+#
+#    Redistribution and use in source and binary forms, with or without
+#    modification, are permitted provided that the following conditions are
+#    met:
+#
+#    1. Redistributions of source code must retain the above copyright notice,
+#       this list of conditions and the following disclaimer.
+#
+#    2. Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#
+#    3. Neither the name of the QuTiP: Quantum Toolbox in Python nor the names
+#       of its contributors may be used to endorse or promote products derived
+#       from this software without specific prior written permission.
+#
+#    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+#    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+#    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+#    PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+#    LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+#    DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+#    THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+#    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+#    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+###############################################################################
+import numpy as np
+
+
+__all__ = ['GateCompiler']
+
+
+class GateCompiler(object):
+    """
+    Base class. It decomposes a :class:`qutip.QubitCircuit` into
+    the pulse sequence for the processor.
+
+    Parameters
+    ----------
+    N: int
+        The number of the component systems.
+
+    params: dict
+        A Python dictionary contains the name and the value of the parameters,
+        such as laser frequency, detuning etc.
+
+    num_ops: int
+        Number of control Hamiltonians in the processor.
+
+    Attributes
+    ----------
+    N: int
+        The number of the component systems.
+
+    params: dict
+        A Python dictionary contains the name and the value of the parameters,
+        such as laser frequency, detuning etc.
+
+    num_ops: int
+        Number of control Hamiltonians in the processor.
+
+    gate_decomps: dict
+        The Python dictionary in the form of {gate_name: decompose_function}.
+        It saves the decomposition scheme for each gate.
+    """
+    def __init__(self, N, params, num_ops):
+        self.gate_decomps = {}
+        self.N = N
+        self.params = params
+        self.num_ops = num_ops
+
+    def decompose(self, gates):
+        """
+        Decompose the the elementary gates
+        into control pulse sequence.
+
+        Parameters
+        ----------
+        gates: list
+            A list of elementary gates that can be implemented in this
+            model. The gate names have to be in `gate_decomps`.
+
+        Returns
+        -------
+        tlist: array_like
+            A NumPy array specifies the time of each coefficient
+
+        coeffs: array_like
+            A 2d NumPy array of the shape ``(len(ctrls), len(tlist))``. Each
+            row corresponds to the control pulse sequence for
+            one Hamiltonian.
+
+        global_phase: bool
+            Recorded change of global phase.
+        """
+        # TODO further improvement can be made here,
+        # e.g. merge single qubit rotation gate, combine XX gates etc.
+        self.dt_list = []
+        self.coeff_list = []
+        for gate in gates:
+            if gate.name not in self.gate_decomps:
+                raise ValueError("Unsupported gate %s" % gate.name)
+            self.gate_decomps[gate.name](gate)
+        coeffs = np.vstack(self.coeff_list).T
+
+        tlist = np.empty(len(self.dt_list))
+        t = 0
+        for i in range(len(self.dt_list)):
+            t += self.dt_list[i]
+            tlist[i] = t
+        return np.hstack([[0], tlist]), coeffs
