+++ qutip/qip/device/modelprocessor.py
@@ -44,7 +44,7 @@ from qutip.qip.circuit import QubitCircuit
 from qutip.qip.device.processor import Processor
 
 
-__all__ = ['ModelProcessor', 'GateDecomposer']
+__all__ = ['ModelProcessor']
 
 
 class ModelProcessor(Processor):
@@ -241,84 +241,3 @@ class ModelProcessor(Processor):
             ax.set_title(title)
         fig.tight_layout()
         return fig, ax
-
-
-class GateDecomposer(object):
-    """
-    Base class. It decomposes a :class:`qutip.QubitCircuit` into
-    the pulse sequence for the processor.
-
-    Parameters
-    ----------
-    N: int
-        The number of the component systems.
-
-    params: dict
-        A Python dictionary contains the name and the value of the parameters,
-        such as laser frequency, detuning etc.
-
-    num_ops: int
-        Number of control Hamiltonians in the processor.
-
-    Attributes
-    ----------
-    N: int
-        The number of the component systems.
-
-    params: dict
-        A Python dictionary contains the name and the value of the parameters,
-        such as laser frequency, detuning etc.
-
-    num_ops: int
-        Number of control Hamiltonians in the processor.
-
-    gate_decomps: dict
-        The Python dictionary in the form of {gate_name: decompose_function}.
-        It saves the decomposition scheme for each gate.
-    """
-    def __init__(self, N, params, num_ops):
-        self.gate_decomps = {}
-        self.N = N
-        self.params = params
-        self.num_ops = num_ops
-
-    def decompose(self, gates):
-        """
-        Decompose the the elementary gates
-        into control pulse sequence.
-
-        Parameters
-        ----------
-        gates: list
-            A list of elementary gates that can be implemented in this
-            model. The gate names have to be in `gate_decomps`.
-
-        Returns
-        -------
-        tlist: array_like
-            A NumPy array specifies the time of each coefficient
-
-        coeffs: array_like
-            A 2d NumPy array of the shape ``(len(ctrls), len(tlist))``. Each
-            row corresponds to the control pulse sequence for
-            one Hamiltonian.
-
-        global_phase: bool
-            Recorded change of global phase.
-        """
-        # TODO further enhancement can be made here,
-        # e.g. merge single qubit rotation gate, combine XX gates etc.
-        self.dt_list = []
-        self.coeff_list = []
-        for gate in gates:
-            if gate.name not in self.gate_decomps:
-                raise ValueError("Unsupported gate %s" % gate.name)
-            self.gate_decomps[gate.name](gate)
-        coeffs = np.vstack(self.coeff_list).T
-
-        tlist = np.empty(len(self.dt_list))
-        t = 0
-        for i in range(len(self.dt_list)):
-            t += self.dt_list[i]
-            tlist[i] = t
-        return np.hstack([[0], tlist]), coeffs
